Update to 2.1pre6:
This update has been a long time in coming, and was spurred by some
recent eBay changes that broke drag'n'drop and searches. Other eBay
changes broke high-bidder recognition, and determining
whether a listing was an auction or a fixed price listing. This
should fix all of those, as well as several UI issues, especially
for Mac OS X Tiger users.

A great feature in the new version is thumbnail mode, which is
essentially a view of your listing that includes the thumbnail on
the same line as your auction information. It's been cleaned up
and made to look very nice. You can configure your columns to add
the Thumbnail column using the small '*' above the scroll bar on
the right, or under the Tab customization menu.

An overview of the features fixed for the new version are:

    * Bugs fixed
	  o Fix the blank-screen (but auctions are there, and only
	  visible when selected) on Mac OS X Tiger
	  o The newest forms of bad HTML found on eBay listings
	  shouldn't break the parser anymore
          o Don't mistake auctions for fixed price listings
          o Recognize eBay's new URL format.
          o Fix drag-and-drop on Linux (patch by Eric)
	  o Don't initiate sniping until the eBay time has been
	  loaded from the server
	  o If the pre-snipe had already fired (2 minutes before
	  the actual snipe), and you changed the snipe amount, it
	  would instantly place a bid for the amount you entered
	  as a snipe. This is fixed.
    * Small features and improvements
	  o Don't present a dialog message when cancelling a
	  multisnipe; events like that should just quietly obey.
	  o Reduce the number of unimportant exceptions that show
	  in the log file
          o Provide the ability to easily submit error logs
          o Improve the integration with My JBidwatcher
          o Improve high bidder recognition substantially
	  o Simplify tab save/restore handling, and apply a tab
	  ordering fix (patch from Andreas Stempfhuber).
	  o Added a patch from mickle_ak which should improve
	  corporate proxy support.
    * Features
          o Initial support for Gixen integration via My JBidwatcher
		+ This isn't complete, but it's getting close
	  o Deep improvements to the Thumbnail view and taking the
	  Mac UI improvements and applying them everywhere, including
	  (I believe) a working implementation on Windows and Linux!